+++ app/src/main/java/com/example/swith/ui/manage/ManageUserActivity.kt
@@ -15,7 +15,7 @@ import com.google.android.material.tabs.TabLayoutMediator
 
 class ManageUserActivity : AppCompatActivity(), View.OnClickListener{
     lateinit var binding : ActivityManageUserBinding
-    var groupIdx : Int = -1
+    var groupIdx : Long = -1
     private val tabTitleArray = arrayOf(
         "지원",
         "목록"
@@ -38,7 +38,7 @@ class ManageUserActivity : AppCompatActivity(), View.OnClickListener{
 
     fun initData()
     {
-        (intent.hasExtra("groupIdx")).let { groupIdx = intent.getIntExtra("groupIdx", 0) }
+        (intent.hasExtra("groupIdx")).let { groupIdx = intent.getLongExtra("groupIdx", 0) }
         Log.e("summer","groupIdx = ${groupIdx}")
     }
